Add tests for failing second get_payload().
